+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Turn on and control logging through the CLI.
 - `colorize` keyword argument for `enable_logging` for logs to not use ANSI color codes.
 
 ### Fixed